Extracted handlers into separate files.
* Added flowercatalog app.
