+++ public/html/ageratum.html
@@ -1,0 +1,26 @@
+<html>
+
+<head>
+  <title>Ageratum</title>
+  <link rel="stylesheet" href="/styles/catalogs.css" />
+</head>
+
+<body>
+  <div class="page-header">
+    <a href="/">Home</a>
+    <h1>Ageratum</h1>
+  </div>
+
+  <div class="page-body">
+    <img src="/images/agerantum.jpg" alt="agerantumImage" />
+    <div class="about-flower">
+      <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Laboriosam, laborum ullam culpa doloribus quod
+        consectetur accusamus consequuntur sit voluptatibus numquam ex nobis deserunt nemo deleniti porro nihil earum
+        cumque voluptates.</p>
+      <a href="/database/Ageratum.pdf" target="_self">Download as book</a>
+    </div>
+  </div>
+
+</body>
+
+</html>
