+++ public/styles/homePage.css
@@ -1,0 +1,20 @@
+body {
+  height: 16px;
+  font-family: sans-serif;
+}
+
+.page-header {
+  border-bottom: 1px solid black;
+  margin-bottom: 10px;
+}
+
+.page-header h1 {
+  color: #ffb221;
+  text-align: center;
+}
+
+.page-body {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+}
